1.分析界面动态添加数据

+++ app/src/main/java/life/chenshi/keepaccounts/bean/SumMoneyByDateAndTypeBean.kt
@@ -4,8 +4,8 @@ import java.math.BigDecimal
 
 class SumMoneyByDateAndTypeBean(private val day: String, private val sumMoney: BigDecimal) {
 
-    fun getDay(): Float {
-        return day.toFloat()
+    fun getDay(): Int {
+        return day.toInt()
     }
 
     fun getMoney(): Float {
